+++ CHANGELOG.md
@@ -82,4 +82,5 @@
 - MVP-408 - Add Market research landing (category) page
 - No-Ticket - Change behaviour of users context export plan to 'GET
 - MVP-425 - Updates to Learn homepage
+- MVP-526 - Add Learn introduction page with slider
 - MVP-433 - Fix last step of signin 404